% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualtrend.R
\name{annualtrend}
\alias{annualtrend}
\alias{annualfit}
\alias{annualtrendmean}
\alias{annualtrendvariance}
\alias{annualtrendquantile}
\alias{annualtrendsd}
\alias{annualtrendquan}
\title{Annual linear trends}
\usage{
annualtrendmean(
  data,
  datamonth,
  datayear,
  showmonths = c(1:12),
  monthly = FALSE,
  limaxisy = c(0, 40),
  title = TRUE,
  linear = T,
  low = F,
  width = 0.2,
  legend = TRUE,
  legpos = "topleft",
  legfont = 0.7,
  legcols = 3,
  english = TRUE
)

annualtrendsd(
  data,
  datamonth,
  datayear,
  showmonths = c(1:12),
  monthly = FALSE,
  desv = 1,
  limaxisy = c(0, 12),
  title = TRUE,
  linear = T,
  low = F,
  width = 0.2,
  legend = TRUE,
  legpos = "topleft",
  legfont = 0.7,
  legcols = 3,
  english = TRUE
)

annualtrendquan(
  data,
  datamonth,
  datayear,
  showmonths = c(1:12),
  monthly = FALSE,
  quan = 0.5,
  limaxisy = c(0, 40),
  title = TRUE,
  linear = T,
  low = F,
  width = 0.2,
  legend = TRUE,
  legpos = "topleft",
  legfont = 0.7,
  legcols = 3,
  english = TRUE
)
}
\arguments{
\item{data}{Data frame or matrix of the data that we are representing (when monthly=TRUE we can only plot one place or column of data).}

\item{datamonth}{Array of the months associated to the data.}

\item{datayear}{Array of the year associated to the data.}

\item{showmonths}{Array of the months that we want to plot when monthly=TRUE.}

\item{monthly}{If monthly=TRUE the plot will represent the annual trend of the specified months (monthly means) of a place. If monthly=FALSE, the trends will fit annual means of each one of the chosen places.}

\item{limaxisy}{The range represented of the vertical axis.}

\item{title}{The title of the plot. If TRUE, the title will have its default value, if FALSE there won't be a title. Otherwise, the title will be the string value of this parameter.}

\item{linear}{Boolean. If TRUE, the plot will show the fitted trends.}

\item{low}{Boolean. If TRUE, the plot will show the fitted lowess.}

\item{width}{Width of the window of data in the lowess fit.}

\item{legend}{The legend of the plot. If TRUE, the legend will have its default value, if FALSE there won't be a legend. Otherwise, the legend will be the array of strings assigned of this parameter.}

\item{legpos}{Position of the legend.}

\item{legfont}{Font size of the legend.}

\item{legcols}{Number of columns of the legend.}

\item{english}{Boolean. If TRUE, the axis labels will be in english. Otherwise, it will be in spanish.}
}
\description{
A plot of the annual linear trends or lowess fits of the data in the places desired or in the chosen months of a specific place.
}
\examples{
x<-c(30,31,33,30,29,28,29,32,35)
data<-data.frame("test"=c(30,31,33,30,29,28,29,32,35),"test2"=c(30,31,33,30,29,28,29,32,35))
y<-rep(c(1,2,3),3)
year<-c(rep(1,3),rep(2,3),rep(3,3))
annualtrend(data=data,datamonth=y,datayear=year,showmonths=c(1,2,3))
annualtrend(data=data[1],datamonth=y,datayear=year,showmonths=c(1,2,3),monthly=TRUE)

aux.df<- read.table( 'Tmax_Aragon.txt', header=T, dec='.'  )
annualtrendmean(aux.df[5:22],aux.df$month,aux.df$year)
annualtrendmean(aux.df[17],aux.df$month,aux.df$year)
annualtrendmean(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE)
annualtrendmean(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,showmonths = c(6,9),limaxisy=c(25,32),english=FALSE,title=FALSE)
annualtrendmean(aux.df[c(17,18)],aux.df$month,aux.df$year,monthly=FALSE,limaxisy=c(18,23),english=FALSE,title=FALSE)

annualtrendsd(aux.df[5:22],aux.df$month,aux.df$year)
annualtrendsd(aux.df[17],aux.df$month,aux.df$year)
annualtrendsd(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,limaxisy=c(3,5))

annualtrendquan(aux.df[5:22],aux.df$month,aux.df$year)
annualtrendquan(aux.df[17],aux.df$month,aux.df$year)
annualtrendquan(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE)
annualtrendquan(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,showmonths = 6,quan=c(0.05,0.5,0.95),limaxisy=c(15,40),english=FALSE,title=FALSE)
annualtrendquan(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,showmonths = c(6,9),quan=c(0.05,0.5,0.95),limaxisy=c(15,45),english=FALSE,title=FALSE,legcols=2)
annualtrendquan(aux.df[c(17,18)],aux.df$month,aux.df$year,monthly=FALSE,quan=c(0.05,0.5,0.95),limaxisy=c(0,50),english=FALSE,title=FALSE,legcols=2)

?Cu?ndo el cuantil 0.9 de junio en zaragoza alcanza el valor original del cuantil 0.95?
table<-annualtrendquan(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,showmonths = 6,quan=c(0.9,0.95),limaxisy=c(15,40))
abline(v=(table[2,1]+table[2,2]*1953-table[1,1])/table[1,2])

?El cuantil 0.5 de enero de 2015 alcanza el quantil 0.5 de febrero de 1953?
table<-annualtrendquan(aux.df[17],aux.df$month,aux.df$year,monthly=TRUE,showmonths = c(1,2),quan=0.5,limaxisy=c(5,20))
(table[2,1]+table[2,2]*1953-table[1,1])/table[1,2]<2015 #No lo alcanza

aux.threshold<-tapply(aux.df$Tx.zaragoza,aux.df$month,quantile,0.50)
aux.marcador<-(aux.df[17]>aux.threshold[aux.df$month])
annualtrendmean(aux.df[5:22][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendmean(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendmean(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],monthly=TRUE)

annualtrendsd(aux.df[5:22][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendsd(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendsd(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],monthly=TRUE,limaxisy = c(1,3))

annualtrendquan(aux.df[5:22][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendquan(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendquan(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],monthly=TRUE)

aux.marcador<-(aux.df$month==1)|(aux.df$month==2)
annualtrendmean(aux.df[5:22][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendmean(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador])
annualtrendmean(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],showmonths=c(1,2),monthly=TRUE)

annualtrendsd(aux.df[5:22][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],limaxis=c(2,6))
annualtrendsd(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],limaxis=c(2,6))
annualtrendsd(aux.df[17][aux.marcador,],aux.df$month[aux.marcador],aux.df$year[aux.marcador],limaxis=c(2,6),showmonths=c(1,2),monthly=TRUE)
}
