% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmeses.R
\name{boxmonths}
\alias{boxmonths}
\alias{boxmeses}
\title{Boxplot by months}
\usage{
boxmonths(data, title = "", english = TRUE)
}
\arguments{
\item{data}{Array of two columns: one with the data that we want to represent and the other with the months linked to each value.}

\item{title}{String that will appear as the title of the boxplot.}

\item{english}{Boolean. If TRUE, the axis labels will be in english. Otherwise, it will be in spanish.}
}
\description{
A boxplot of the data separated in months.
}
\examples{
x<-c(30,31,33,30,29,28,29,32,35)
y<-rep(c(1,2,3),3)
boxmonths(cbind(x,y))

data<- read.table( 'Tmax_Aragon.txt', header=T, dec='.'  )
aux.data<-cbind(data$Tx.zaragoza,data$month)
boxmonths(aux.data,english=FALSE)

aux.threshold<-tapply(data$Tx.zaragoza,data$month,quantile,0.50)
aux.marcador<-(data$Tx.zaragoza>aux.threshold[data$month])
boxmonths(aux.data[aux.marcador],english=FALSE)

anomaly<-(data$Tx.zaragoza-aux.threshold[data$month])
aux.data<-cbind(anomaly,data$month)
boxmonths(aux.data,english=FALSE)
}
