% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthcorrelation.R
\name{persistence}
\alias{persistence}
\alias{mescorrelation}
\alias{monthcorrelation}
\title{Autocorrelation by months}
\usage{
monthcorrelation(data, dates, lags = 1, monthly = TRUE)
}
\arguments{
\item{data}{Array of the data.}

\item{dates}{Data frame of dates asociated to the data with columns for the year, the day of the year (day.year) and the month.}

\item{monthly}{If monthly=TRUE the correlation will be calculated by months. If monthly=FALSE,the function will calculate the same autocorrelation coefficients that can be aobtained by the function acf.}

\item{lag}{Number of days of lag that we want to obtain the correlation coefficients for.}
}
\description{
A matrix of the autocorrelation coefficients of the data that can be separated by months.
}
\examples{
data<-c(30,31,33,30,29,28,29,32,35)
dates<-data.frame(row.names=1:9)
dates$year<-rep(1999,9)
dates$month<-rep(1,9)
dates$day.year<-1:9
monthcorrelation(data,dates)

auxiliar<- read.table( 'Tmax_Aragon.txt', header=T, dec='.'  )
cormatrix<-monthcorrelation(auxiliar[,places[13]],auxiliar[,1:4],lags=c(1:5))
monthcorrelation(auxiliar[,places[13]],auxiliar[,1:4],lags=c(1:5),monthly=FALSE)
monthcorrelation(auxiliar[,places[13]],auxiliar[,1:4],lags=c(1:5),monthly=TRUE)
acf(auxiliar[,places[13]])[1:5]
}
