% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preselection.R
\name{preselection}
\alias{preselection}
\alias{preselecion}
\alias{premodel}
\alias{localfits}
\alias{ajuste.parada}
\title{Algorithm of preselection of linear models}
\usage{
ajuste.parada(
  place,
  data = auxiliar,
  pendiente = FALSE,
  rstanarm = FALSE,
  bamlss = FALSE,
  verbal = TRUE
)
}
\arguments{
\item{place}{String. Name of the place whose data we are modeling: "Tx.(place)".}

\item{data}{List of data. The data set must have the next columns:

-(place): Data of the maximum daily temperature of a specific place.

-sin1,cos1,sin2,cos2,sin3,cos3,sin4,cos4,sin5,cos5,sin6,cos6: Daily values of the harmonic functions of periods 1 year, 1/2 year, 1/3 year, 1/4 year, 1/5 year and 1/6 year.

-year: Year of each temperature.

-day.year: Day of the year of each temperature.}

\item{pendiente}{Boolean. If TRUE the annual trend of the temperature and its interaction with the harmonic functions will be plotted.}

\item{rstanarm}{Boolean. If TRUE a model Bayesian model will be fitted using the function stan_glm and the parameters of the best linear model.}

\item{bamlss}{Boolean. If TRUE a model Bayesian model will be fitted using the function bamlss and the parameters of the best linear model.}

\item{verbal}{Boolean. If TRUE the function will print the results.}
}
\value{
A list of the five models fitted and the answer to which one of them is the best.
}
\description{
Five different linear models are fitted and compared to represent the maximum daily temperature data.
}
\examples{
auxiliar<- read.table( 'Tmax_Aragon.txt', header=T, dec='.'  )
auxiliar$sin1 <- sin( (auxiliar$day.year-1)/366  *pi *2 )
auxiliar$cos1 <- cos( (auxiliar$day.year-1)/366  *pi *2 )
auxiliar$sin2 <- sin( (auxiliar$day.year-1)/366  *pi *4 )
auxiliar$cos2 <- cos( (auxiliar$day.year-1)/366  *pi *4 )
auxiliar$sin3 <- sin( (auxiliar$day.year-1)/366  *pi *6 )
auxiliar$cos3 <- cos( (auxiliar$day.year-1)/366  *pi *6 )
auxiliar$sin4 <- sin( (auxiliar$day.year-1)/366  *pi *8 )
auxiliar$cos4 <- cos( (auxiliar$day.year-1)/366  *pi *8 )
auxiliar$sin5 <- sin( (auxiliar$day.year-1)/366  *pi *10 )
auxiliar$cos5 <- cos( (auxiliar$day.year-1)/366  *pi *10 )
auxiliar$sin6 <- sin( (auxiliar$day.year-1)/366  *pi *12 )
auxiliar$cos6 <- cos( (auxiliar$day.year-1)/366  *pi *12 )
preselection("Tx.zaragoza")
}
