% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleqqplot.R
\name{sampleqqplot}
\alias{sampleqqplot}
\title{Q-Q plot of the quantile distributions}
\usage{
sampleqqplot(
  marginquantilematrix,
  empiricquantilevector,
  qqplot = TRUE,
  range = 1,
  xlab = "Empiric margin quantiles",
  ylab = "Simulated margin quantiles"
)
}
\arguments{
\item{marginquantilematrix}{List with the arrays of samples for each marginal quantile.}

\item{empiricquantilevector}{List of the empiric marginal quantiles.}

\item{qqplot}{Boolean. If TRUE the function will represent the q-q plot. Otherwise it will just calculate its important summary information.}

\item{range}{Radius of the interval around the y=x (sampled=empirical)function where we consider the sampled marginal quantiles as correct. Its default value is 1.}

\item{xlab, ylab}{Axis labels of the plot. Their default values are "Empiric margin quantiles" and "Simulated margin quantiles", respectively.}
}
\description{
Using the empiric and sampled quantiles, prepare a Q-Q plot with the chosen simple credibility interval and calculate the bias.
}
